+++ src/ruiz/ssh/service/MaterialService.java
@@ -13,6 +13,6 @@ public interface MaterialService {
     public List<Material> queryAll();
     public List<Material> match(String s);
     
-    public Material tbadd(String name, String type, String detail, int weight, int volume);
-	public boolean tbupdate(int id, String name, String type, String detail, int weight, int volume);
+    public Material tbadd(String name, String type, String detail, double weight, double volume);
+	public boolean tbupdate(int id, String name, String type, String detail, double weight, double volume);
 }
